requirements.txt
pandas==2.2.3
protobuf==5.29.3
python-dotenv==1.0.1
streamlit==1.41.1
